Move to dto.transaction
